refactor: domain 패키지 하위 클래스들의 패키지 상세화
상위 객체임이 명확히 드러날 수 있도록 패키지 구조 정비.
