+++ src/main/java/edu/nextstep/racing/domain/game/cars/car/Name.java
@@ -1,4 +1,4 @@
-package edu.nextstep.racing.domain;
+package edu.nextstep.racing.domain.game.cars.car;
 
 import java.util.Objects;
