+++ src/test/java/edu/nextstep/racing/domain/NameTest.java
@@ -5,6 +5,8 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import edu.nextstep.racing.domain.game.cars.car.Name;
+
 @DisplayName("Location : 자동차의 이름에 대한 원시값 포장 클래스")
 class NameTest {
